Adds rbac to give argocd admin permission to the namespace (#28)

+++ .github/scripts/validate-deploy.sh
@@ -38,5 +38,13 @@ fi
 echo "Printing payload/1-infrastructure/namespace/${NAMESPACE}/namespace/ns.yaml"
 cat "payload/1-infrastructure/namespace/${NAMESPACE}/namespace/ns.yaml"
 
+if [[ ! -f "payload/1-infrastructure/namespace/${NAMESPACE}/namespace/rbac.yaml" ]]; then
+  echo "Payload missing: payload/1-infrastructure/namespace/${NAMESPACE}/namespace/rbac.yaml"
+  exit 1
+fi
+
+echo "Printing payload/1-infrastructure/namespace/${NAMESPACE}/namespace/rbac.yaml"
+cat "payload/1-infrastructure/namespace/${NAMESPACE}/namespace/rbac.yaml"
+
 cd ..
 rm -rf .testrepo
